discouple IO submdoules e4c44e4

+++ Microsoft.VisualBasic.Core/ComponentModel/Algorithm/DynamicProgramming.vb
@@ -41,6 +41,6 @@
 
 Namespace ComponentModel.Algorithm.DynamicProgramming
 
-    Public Delegate Function ISimilarity(Of T)(x As T, y As T) As Integer
+    Public Delegate Function ISimilarity(Of T)(x As T, y As T) As Double
 
 End Namespace
